Imported and installed uuidv4 imported types "DraftExpense and Expense"
added the budgetaction type add-expense ...darftexpense

added in the staate local with a empty Expense the same in the initialState

created a  create DraftExpense  with copy draft and return generated id. with uuidv4

created an action for add expense and take a copy of state, state.expense and expense

added modal: false. to closed when you make a click on add new expense in the modal form